% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V2StachUtilities.R
\docType{class}
\name{V2StachUtilities}
\alias{V2StachUtilities}
\title{V2StachUtilities}
\description{
The purpose of this class is to provide the helper methods
}
\examples{
\dontrun{
package <- 'Stach data which is converted into RowOrganized Package'
tableData <-
V2StachUtilities$public_methods$GetMappingObject(package$tables,
primaryTableId)$value
}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GetMappingObject}{\code{V2StachUtilities$GetMappingObject()}}
\item \href{#method-ValueToString}{\code{V2StachUtilities$ValueToString()}}
\item \href{#method-clone}{\code{V2StachUtilities$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetMappingObject"></a>}}
\if{latex}{\out{\hypertarget{method-GetMappingObject}{}}}
\subsection{Method \code{GetMappingObject()}}{
Helper method for getting the value corresponding to the given key from the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V2StachUtilities$GetMappingObject(objectMap, key, keyname = "key")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{objectMap}}{Mapping object value}

\item{\code{key}}{Mapping object corresponding key value}

\item{\code{keyname}}{Mapping object corresponding key name}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Returns the value corresponding to the given key from the object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ValueToString"></a>}}
\if{latex}{\out{\hypertarget{method-ValueToString}{}}}
\subsection{Method \code{ValueToString()}}{
Helper method for converting value to string
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V2StachUtilities$ValueToString(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{Value that should be converted to string}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Returns the value by converting corresponding value to string
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V2StachUtilities$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
