% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V2ColumnOrganizedStachUtilities.R
\docType{class}
\name{V2ColumnOrganizedStachUtilities}
\alias{V2ColumnOrganizedStachUtilities}
\title{V2ColumnOrganizedStachUtilities}
\description{
Provide helper functions for column organized stach
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-V2ColumnOrganizedStachUtilities-GetPrimaryTableIds}{\code{V2ColumnOrganizedStachUtilities$GetPrimaryTableIds()}}
\item \href{#method-V2ColumnOrganizedStachUtilities-DecompressAll}{\code{V2ColumnOrganizedStachUtilities$DecompressAll()}}
\item \href{#method-V2ColumnOrganizedStachUtilities-clone}{\code{V2ColumnOrganizedStachUtilities$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V2ColumnOrganizedStachUtilities-GetPrimaryTableIds"></a>}}
\if{latex}{\out{\hypertarget{method-V2ColumnOrganizedStachUtilities-GetPrimaryTableIds}{}}}
\subsection{Method \code{GetPrimaryTableIds()}}{
Get Primary Table Ids
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V2ColumnOrganizedStachUtilities$GetPrimaryTableIds(package)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{package}}{Stach Data which is represented as a Package object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Returns list of primary table ids
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V2ColumnOrganizedStachUtilities-DecompressAll"></a>}}
\if{latex}{\out{\hypertarget{method-V2ColumnOrganizedStachUtilities-DecompressAll}{}}}
\subsection{Method \code{DecompressAll()}}{
Get decompress stach data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V2ColumnOrganizedStachUtilities$DecompressAll(package)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{package}}{Stach Data which is represented as a Package object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Returns decompressed stach data
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-V2ColumnOrganizedStachUtilities-clone"></a>}}
\if{latex}{\out{\hypertarget{method-V2ColumnOrganizedStachUtilities-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{V2ColumnOrganizedStachUtilities$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
